Cleanup and refactoring of foo

+++ src/foo/bar.cpp
@@ -1,8 +1,6 @@
 #include "bar.h"
 
-// using namespace foo;
-
 // IMPL
 namespace foo {
-int bar() { return 42; }
+uint16_t bar() { return 42; }
 } // namespace foo
