+++ src/foo/bar.h
@@ -1,8 +1,10 @@
 #ifndef FOO_H
 #define FOO_H
 
+#include <cstdint>
+
 namespace foo {
-int bar();
+uint16_t bar();
 } // namespace foo
 
 #endif // FOO_H
